ミッション→カテゴリ リレーション追加

+++ database/migrations/2020_07_15_085710_create_missions_table.php
@@ -15,8 +15,8 @@ class CreateMissionsTable extends Migration
     {
         Schema::create('missions', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->unsigned();
-            $table->integer('category_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
+            $table->bigInteger('category_id')->unsigned();
             $table->string('name', 20);
             $table->string('level_unit', 10)->default('回');
             $table->string('memo', 100)->nullable();
